Revert "Remove "Steam Deck Logo 14" wallpaper"
Apparently it is intended.

From a conversation with Greg.
"Re: wallpaper 54 - it's sort of a terrible-on-purpose joke - we were thinking
of renaming that image troll.jpg or something? The image was originally made for
a kind of video commercial that we're making, and when someone had the idea of
turning these images into desktop wallpapers we thought it would be funny to
include this one."