Use hicolor icon theme for kickoff icon
Instead of setting the icon of kickoff with a direct path to a PNG
(which looked a bit jagged), this uses the hicolor icon theme to set the
kickoff icon. Previously, switching to Breeze Light would make the
kickoff icon difficult to see. This icon also supports recoloring, so if
anyone wanted to make a Vapor Light theme, they wouldn't need to change
the icon at all.